+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## [Unreleased]
+* 57261: Added attribute helper
 
 ## [1.0.0] 2019-12-06